requirements.txt
tensorflow>=2.12,<2.17
protobuf<4.0.0
numpy
pillow
matplotlib
opencv-python
absl-py
pycocotools-windows
lxml
Cython
pandas
scikit-learn
easyocr
